Refactor. Removing Why, WhatIf and How
