+++ app/components/Question/styles.css
@@ -3,7 +3,7 @@
     display: flex;
     flex-direction: column;
     border: 2px solid #032760;
-    /*padding: 20px 10px 10px 30px;*/
+    padding: 20px 10px 10px 30px;
     background: #ffeffc;
     margin: 5px;
     border-radius: 25px;
@@ -38,20 +38,23 @@
 .topRow {
 
 }
-.icon {
-    /*flex: 1;*/
-}
 
+.thinkingIcon {
+    cursor: pointer;
+    color: red;
+}
 .askQuestionIcon {
     float: right;
 }
 
-icon {
+.icon {
     cursor: pointer;
     marginRight: 10
 }
 
-starred {
+.starred {
+    cursor: pointer;
+    marginRight: 10;
     color: gold;
 }
