--- app/components/Questions/QuestionView/QuestionOld/question.css
@@ -1,47 +1,0 @@
-.questionBox {
-
-    display: flex;
-    flex-direction: column;
-    border: 2px solid #032760;
-    /*padding: 20px 10px 10px 30px;*/
-    background: #ffeffc;
-    margin: 5px;
-    border-radius: 25px;
-    font-size: 30px;
-    /*margin-left: 50px;*/
-    width: 100%;
-}
-
-.questionText {
-    flex: 9;
-    /*padding: 22px;*/
-    padding-left: 10px;
-    text-decoration: none;
-
-}
-
-
-.askQuestion {
-    border-left: 1px solid black;
-    margin: 10px;
-    padding-left: 5px;
-}
-
-
-.iconRow {
-    /*justify-self: flex-end;*/
-    display: inline-block;
-    /*flex-direction: row;*/
-    padding-left: 10px;
-    width: 100%;
-}
-
-.icon {
-    /*flex: 1;*/
-}
-
-.questionIcon {
-    float: right;
-    display: block;
-}
-
