--- app/components/Questions/Why.js
@@ -1,84 +1,0 @@
-import React from "react";
-import { connect } from "react-redux";
-import { compose, graphql } from "react-apollo";
-import update from 'immutability-helper';
-
-import Notifications from 'react-notification-system-redux';
-import { unauthorizedErrorNotification } from '../../notifications/error.notifications';
-
-import QuestionViewContainer from "./QuestionView/QuestionViewContainer";
-import { updateCurrentWhy } from "../../actions/Why";
-import { setQuestionType, WHY, WHATIF } from '../../actions/QuestionPopup';
-
-import QUESTIONS_QUERY from "../../graphql/querys/questions.query";
-
-
-
-const mapDispatchToProps = function (dispatch) {
-  return {
-    onSelectQuestion: (why) => {
-      dispatch(setQuestionType(WHATIF))
-      dispatch(updateCurrentWhy(why))
-    },
-    setQuestionType: () => {
-      dispatch(setQuestionType(WHY))
-    },
-    unAuthorized: () => {
-      console.log("DISPATCH UNAUTHORIZED")
-      dispatch(Notifications.error(unauthorizedErrorNotification))
-    }
-  }
-};
-
-const ITEMS_PER_PAGE = 5;
-
-const Why = compose(
-  connect(
-    null,
-    mapDispatchToProps,
-  ),
-
-  graphql(QUESTIONS_QUERY, {
-    options: (props) => ({
-      variables: {first: ITEMS_PER_PAGE},
-      // pollInterval: 5000
-    }),
-    props: ({ ownProps, data: { fetchMore, loading, error, questions } }) => ({
-      loading,
-      error,
-      connection: questions,
-      onSelectQuestion: ownProps.onSelectQuestion,
-      link: "/whatif",
-      refetchQuery: QUESTIONS_QUERY,
-      currentWhy: null,
-      currentWhatIf: null,
-      // currentUser: ownProps.currentUser,
-      loadMoreEntries() {
-        fetchMore({
-          variables: {
-            after: questions.edges[questions.edges.length - 1].cursor,
-
-          },
-          updateQuery: (previousResult, { fetchMoreResult }) => {
-            // we will make an extra call to check if no more entries
-            if (!fetchMoreResult) {
-              return previousResult;
-            }
-            // push results (older whys) to end of whys list
-            return update(previousResult, {
-
-              questions: {
-                  edges: { $push: fetchMoreResult.questions.edges },
-                  pageInfo: { $set: fetchMoreResult.questions.pageInfo },
-                },
-
-            });
-          }
-        })
-      }
-    })
-  }),
-
-)(QuestionViewContainer);
-
-export default Why;
