+++ app/components/UserProfile/UserQuestions.js
@@ -4,9 +4,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import UserWhysList from './QuestionLists/UserWhysList';
-import UserWhatIfsList from './QuestionLists/UserWhatIfsList';
-import UserHowsList from './QuestionLists/UserHowsList';
 import UserQuestionsList from './QuestionLists/UserQuestionsList';
 import StaredQuestionsList from './QuestionLists/StaredQuestionsList';
 
@@ -19,18 +16,6 @@ const UserQuestions = (props) => {
   if (props.questionView === LATEST) {
     questionLst = <UserQuestionsList userId={props.userId}/>
   }
-  if (props.questionView === WHY ) {
-    questionLst =  <UserWhysList userId={props.userId}/>
-
-  }
-  if (props.questionView === WHATIF) {
-    questionLst =  <UserWhatIfsList userId={props.userId}/>
-
-  }
-  if (props.questionView === HOW) {
-    questionLst =  <UserHowsList userId={props.userId}/>
-
-  }
   if (props.questionView === STAR) {
     questionLst =  <StaredQuestionsList userId={props.userId}/>
 
@@ -39,9 +24,6 @@ const UserQuestions = (props) => {
     <div>
       <button onClick={() => {props.onViewChange(LATEST)}}>Latest</button>
 
-      <button onClick={() => {props.onViewChange(WHY)}}>Whys</button>
-      <button onClick={() => {props.onViewChange(WHATIF)}}>WhatIfs</button>
-      <button onClick={() => {props.onViewChange(HOW)}}>Hows</button>
 
       <button onClick={() => {props.onViewChange(STAR)}}>Stared Questions</button>
