+++ app/config/routes.js
@@ -3,10 +3,7 @@ import {Router, Route, browserHistory, IndexRoute} from "react-router";
 
 import Main from "../components/Main/Main";
 import Home from '../components/Home/Home';
-import QuestionsView from '../components/Questions/QuestionsView'
-import Why from "../components/Questions/Why";
-import WhatIf from "../components/Questions/WhatIf";
-import How from "../components/Questions/How";
+import QuestionsView from '../components/QuestionViews/QuestionsView'
 import LoginContainer from "../components/Authentication/LoginContainer";
 import RegisterContainer from "../components/Authentication/RegisterContainer";
 import UserProfileContainer from "../components/UserProfile/UserProfileContainer";
@@ -18,12 +15,9 @@ var AppRouter = function () {
         <Route component={Home}>
           <IndexRoute component={QuestionsView}/>
           {/*<Route path="/thread/:name" component={ThreadPage}/>*/}
-          <Route path="popular" component={Why}/>
         </Route>
         <Route path='question/:id' component={QuestionsView}/>
 
-        <Route path='/whatif' component={WhatIf}/>
-        <Route path='/how' component={How}/>
         < Route
           path='/login'
           component={LoginContainer}/>
