--- app/graphql/mutations/createHow.mutation.js
@@ -1,16 +1,0 @@
-import {gql} from 'react-apollo';
-import QUESTION_FRAGMENT from '../fragments/question.fragment';
-
-const CREATE_HOW_MUTATION = gql`
-  mutation createHow($whatIfId: ID!, $question: String!) {
-    createHow(whatIfId: $whatIfId, question: $question) {
-      id
-      question {
-        ... QuestionFragment
-      }
-    }
-  }
-  ${QUESTION_FRAGMENT}
-`;
-
-export default CREATE_HOW_MUTATION;
