+++ app/graphql/mutations/unstarQuestion.mutation.js
@@ -7,7 +7,7 @@ const UNSTAR_QUESTION_MUTATION = gql`
   mutation unstarQuestion($id: ID!) {
     unstarQuestion(id: $id){
        id
-       question
+       questionText
        starredByCurrentUser
        stars
      }
