+++ package.json
@@ -30,6 +30,7 @@
     "express-session": "^1.15.2",
     "faker": "^4.1.0",
     "file-loader": "^0.11.1",
+    "graphql-tag": "^2.8.0",
     "immutability-helper": "^2.5.1",
     "jwt-decode": "^2.2.0",
     "loadash": "^0.0.1",
@@ -38,7 +39,7 @@
     "pg-hstore": "^2.3.2",
     "prop-types": "^15.6.0",
     "react": "^15.4.2",
-    "react-apollo": "^1.0.0",
+    "react-apollo": "^1.4.16",
     "react-bootstrap": "^0.30.8",
     "react-dom": "^15.4.2",
     "react-icons": "^2.2.3",
