+++ server/package.json
@@ -5,12 +5,14 @@
   "main": "server.js",
   "scripts": {
     "start": "nodemon ./server.js --host 0.0.0.0 --exec babel-node -e js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "deploy": "pm2 deploy ecosystem.config.js production"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "babel-cli": "^6.24.0",
+    "babel-core": "^6.26.0",
     "bcrypt": "^1.0.2",
     "body-parser": "^1.17.1",
     "cors": "^2.8.1",
